Optimization with Newton-Raphson Method
This Python script is an example of finding the root (zero) of a mathematical function using the Newton-Raphson method. Below is a brief summary:

This script performs the process of finding the root of a mathematical function by applying the Newton-Raphson method. A new prediction is made at each iteration using the first derivative and second derivative. In an iterative process, starting from the initial value, a certain number of iterations is reached and the zero point of the function is approached.